feat: renamed to private module
