+++ sqlspec/statement/builder/mixins/_cte_and_set_ops.py
