+++ sqlspec/statement/builder/mixins/_pivot_operations.py
